fixed main issue.
code needs rebase to be made fully reduendent.

stupid system does not need to track the number for the system messages.

and will only ever recieve a result and the exact next sequence it expects. so a single thread must only be allowed to send messages and everything else goes through it. hello delay!!